Move error to own file
